net: Fix ttl range

+++ src/kyron/src/net/tcp_listener.rs
@@ -74,7 +74,7 @@ impl TcpListener {
 
     /// Sets the value for the IP_TTL option on this socket.
     /// This value sets the time-to-live field that is used in every packet sent from this socket.
-    pub fn set_ttl(&self, ttl: u32) -> NetResult<()> {
+    pub fn set_ttl(&self, ttl: u8) -> NetResult<()> {
         self.listener.set_ttl(ttl)
     }
 }
